+++ doc/source/releasenotes.rst
@@ -6,6 +6,7 @@ Release notes
 
 0.11.1
 =======
+* Fixed a bug in the ``FixedIndices`` constraint class.
 
 0.11.0
 =======